Simplify tx handling to fix issue #6
It turns out that mysql really doesn't like it if external_lock()
returns an error code like that ones we could have thrown when an
auto-commit transaction was committed. Since we always register a new tx
on every entry to external_lock() (thus ensuring our commit function is
called) there is no need to do a commit in the unlock phase of
external_lock. This seems to fix the issue with failing auto-commits
causing the server to crash